+++ atomicshop/ip_addresses.py
@@ -22,3 +22,19 @@ def is_ip_address(string_value: str, ip_type: Union[Literal['ipv4', 'ipv6'], Non
         return False
 
     return True
+
+
+def get_ip_type(ip_address: str) -> Union[Literal['ipv4', 'ipv6'], None]:
+    """
+    The function checks if the string is an IPv4 or IPv6 address.
+
+    :param ip_address: string to check.
+    :return: string, 'ipv4' or 'ipv6'.
+    """
+
+    if is_ip_address(ip_address, ip_type='ipv4'):
+        return 'ipv4'
+    elif is_ip_address(ip_address, ip_type='ipv6'):
+        return 'ipv6'
+    else:
+        return None